Объединил репозитории по сохранению и получению фильтров в один FilterStorageRepository
