+++ app/src/main/java/ru/practicum/android/diploma/di/RepositoryModule.kt
@@ -12,18 +12,8 @@ import ru.practicum.android.diploma.favorites.data.repository.FavoritesRepositor
 import ru.practicum.android.diploma.favorites.domain.impl.LocalRepository
 import ru.practicum.android.diploma.filter.data.repository.FilterDictionariesRepositoryHHNetworkClientBased
 import ru.practicum.android.diploma.filter.domain.impl.FilterDictionariesRepository
-import ru.practicum.android.diploma.filter.data.repository.AllFilterParametersRepositoryImpl
-import ru.practicum.android.diploma.filter.data.repository.AreaFilterRepositoryImpl
-import ru.practicum.android.diploma.filter.data.repository.CountryFilterRepositoryImpl
-import ru.practicum.android.diploma.filter.data.repository.ExpectedSalaryRepositoryImpl
-import ru.practicum.android.diploma.filter.data.repository.HideNoSalaryItemsRepositoryImpl
-import ru.practicum.android.diploma.filter.data.repository.IndustryRepositoryImpl
-import ru.practicum.android.diploma.filter.domain.impl.AllFilterParameterRepository
-import ru.practicum.android.diploma.filter.domain.impl.AreaFilterRepository
-import ru.practicum.android.diploma.filter.domain.impl.CountryFilterRepository
-import ru.practicum.android.diploma.filter.domain.impl.ExpectedSalaryRepository
-import ru.practicum.android.diploma.filter.domain.impl.HideNoSalaryItemsRepository
-import ru.practicum.android.diploma.filter.domain.impl.IndustryRepository
+import ru.practicum.android.diploma.filter.data.repository.FilterStorageRepositoryImpl
+import ru.practicum.android.diploma.filter.domain.impl.FilterStorageRepository
 import ru.practicum.android.diploma.network.data.HeadHunterRepository
 import ru.practicum.android.diploma.search.data.repository.SearchRepository
 
@@ -32,8 +22,8 @@ val repositoryModule = module {
         NavigatorRepositoryImpl(externalNavigator = get())
     }
 
-    single<AllFilterParameterRepository> {
-        AllFilterParametersRepositoryImpl(filterStorage = get())
+    single<FilterStorageRepository> {
+        FilterStorageRepositoryImpl(filterStorage = get())
     }
 
     factory<SearchRepository> {
@@ -56,25 +46,5 @@ val repositoryModule = module {
 
     factory<FilterDictionariesRepository> {
         FilterDictionariesRepositoryHHNetworkClientBased(client = get(), context = androidContext())
-    }
-
-    single<AreaFilterRepository> {
-        AreaFilterRepositoryImpl(filterStorage = get())
-    }
-
-    single<CountryFilterRepository> {
-        CountryFilterRepositoryImpl(filterStorage = get())
-    }
-
-    single<ExpectedSalaryRepository> {
-        ExpectedSalaryRepositoryImpl(filterStorage = get())
-    }
-
-    single<HideNoSalaryItemsRepository> {
-        HideNoSalaryItemsRepositoryImpl(filterStorage = get())
-    }
-
-    single<IndustryRepository> {
-        IndustryRepositoryImpl(filterStorage = get())
     }
 }
