--- app/src/main/java/ru/practicum/android/diploma/filter/domain/impl/ExpectedSalaryRepository.kt
@@ -1,9 +1,0 @@
-package ru.practicum.android.diploma.filter.domain.impl
-
-interface ExpectedSalaryRepository {
-    fun getExpectedSalary(): String
-
-    fun saveExpectedSalary(salaryAmount: String)
-
-    fun clearExpectedSalary()
-}
