--- app/src/main/java/ru/practicum/android/diploma/filter/domain/impl/HideNoSalaryItemsRepository.kt
@@ -1,7 +1,0 @@
-package ru.practicum.android.diploma.filter.domain.impl
-
-interface HideNoSalaryItemsRepository {
-    fun getHideNoSalaryItems(): Boolean
-
-    fun saveHideNoSalaryItems(hideNoSalaryItems: Boolean)
-}
